ref: split models and readers
